Lots of cleanup, almost there
