+++ zavod/zavod/helpers/__init__.py
@@ -40,7 +40,7 @@ from zavod.helpers.text import clean_note, is_empty, remove_bracketed
 from zavod.helpers.text import multi_split
 from zavod.helpers.sanctions import make_sanction
 from zavod.helpers.addresses import make_address, apply_address, format_address
-from zavod.helpers.dates import extract_years, parse_date, check_no_year, backdate
+from zavod.helpers.dates import extract_years, parse_date, check_no_year
 from zavod.helpers.dates import parse_formats
 from zavod.helpers.identification import make_identification
 from zavod.helpers.excel import convert_excel_cell
@@ -57,7 +57,6 @@ __all__ = [
     "make_identification",
     "extract_years",
     "parse_date",
-    "backdate",
     "parse_formats",
     "check_no_year",
     "convert_excel_cell",
